+++ src/async.h
@@ -101,13 +101,13 @@ class RgbRegionsPercent : public GrayRegionsPercent {
 
 class GrayAllBounds : public Napi::AsyncWorker {
     public:
-        GrayAllBounds(const std::string target, const uint_fast16_t width, const uint_fast16_t height, const uint_fast32_t pixCount, const uint_fast8_t pixDiff, const uint_fast8_t diffsPerc, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
+        GrayAllBounds(const std::string target, const uint_fast32_t width, const uint_fast32_t height, const uint_fast32_t pixCount, const uint_fast8_t pixDiff, const uint_fast8_t diffsPerc, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
         void Execute();
         void OnOK();
     protected:
         const std::string target_;
-        const uint_fast16_t width_;
-        const uint_fast16_t height_;
+        const uint_fast32_t width_;
+        const uint_fast32_t height_;
         const uint_fast32_t pixCount_;
         const uint_fast8_t pixDiff_;
         const uint_fast8_t diffsPerc_;
@@ -120,7 +120,7 @@ class GrayAllBounds : public Napi::AsyncWorker {
 
 class GrayMaskBounds : public GrayAllBounds {
     public:
-        GrayMaskBounds(const std::string target, const uint_fast16_t width, const uint_fast16_t height, const uint_fast32_t pixCount, const uint_fast8_t pixDiff, const uint_fast8_t diffsPerc, const uint_fast32_t bitsetCount, const std::vector<bool> &bitsetVec, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
+        GrayMaskBounds(const std::string target, const uint_fast32_t width, const uint_fast32_t height, const uint_fast32_t pixCount, const uint_fast8_t pixDiff, const uint_fast8_t diffsPerc, const uint_fast32_t bitsetCount, const std::vector<bool> &bitsetVec, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
         void Execute();
     private:
         const uint_fast32_t bitsetCount_;
@@ -131,12 +131,12 @@ class GrayMaskBounds : public GrayAllBounds {
 
 class GrayRegionsBounds : public Napi::AsyncWorker {
     public:
-        GrayRegionsBounds(const uint_fast16_t width, const uint_fast16_t height, const uint_fast32_t pixCount, const uint_fast8_t minDiff, const uint_fast8_t regionsLen, const std::vector<Engine::Region> &regionsVec, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
+        GrayRegionsBounds(const uint_fast32_t width, const uint_fast32_t height, const uint_fast32_t pixCount, const uint_fast8_t minDiff, const uint_fast8_t regionsLen, const std::vector<Engine::Region> &regionsVec, const uint_fast8_t *buf0, const uint_fast8_t *buf1, const Napi::Function &cb);
         void Execute();
         void OnOK();
     private:
-        const uint_fast16_t width_;
-        const uint_fast16_t height_;
+        const uint_fast32_t width_;
+        const uint_fast32_t height_;
         const uint_fast32_t pixCount_;
         const uint_fast8_t minDiff_;
         const uint_fast8_t regionsLen_;
